cleanup services and models to shared
