+++ client/src/app/shared/cable.service.spec.ts
