+++ client/src/app/shared/models/gridwidget.model.ts
@@ -1,0 +1,8 @@
+export interface GridWidget {
+    id: number;
+    position: number;
+    length: number;
+    width: number;
+    grid_id: number;
+    widget_id: number;
+  }
