+++ client/src/app/shared/models/layout.model.ts
@@ -1,0 +1,11 @@
+import { Grid } from "./grid.model";
+import { LayoutGrid } from "./layoutgrid.model";
+
+export interface Layout {
+    id: number;
+    name: string;
+    background: string;
+    duration: number;
+    grids: Grid[];
+    layout_grids: LayoutGrid[];
+  }
