+++ client/src/app/shared/models/layoutgrid.model.ts
@@ -1,0 +1,6 @@
+export interface LayoutGrid {
+    id: number;
+    position: number;
+    layout_id: number;
+    grid_id: number;
+  }
